Fill detailed results when a detail view is loaded
