+++ src/js/views/index.js
@@ -1,7 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import Dashboard from "../components/dashboard.jsx";
-import RecordsList from "../components/recordslist.jsx";
 
 let get_wcag_url = require( "../lib/wcagurl.js" );
 
@@ -82,52 +81,6 @@ class Index {
 						container.append( "<div class='result'><span class='count'>" + buckets[ i ].doc_count + "</span><span class='" + type + "' data-code='" + code_selector + "'>" + code_name + "</span></div>" );
 					}
 				}
-			},
-			error: function( jqXHR, textStatus, errorThrown ) {
-				console.log( jqXHR, textStatus, errorThrown );
-			}
-		} );
-	}
-
-	fillDetails( type, selection ) {
-		if ( "selector" === type ) {
-			var selection_text = document.createElement( "textarea" );
-			selection_text.innerHTML = selection;
-			selection = selection_text.value;
-		}
-
-		var body = {
-			"size": 50,
-			"query": {
-				"bool": {
-					"must": [
-						{ "term": { "typeCode": 1 } }
-					]
-				}
-			}
-		};
-
-		var term = { "term": {} };
-		term.term[ type ] = selection;
-		body.query.bool.must.push( term );
-
-		$( ".result-title" ).html( "<h2>" + selection + "</h2>" );
-		$( ".result-details" ).html( "" );
-
-		$.ajax( {
-			url: "https://public.elastic.wsu.edu/a11y-scan/record/_search",
-			type: "POST",
-			crossDomain: true,
-			dataType: "json",
-			data: JSON.stringify( body ),
-			success: function( response ) {
-				let results = [];
-
-				for ( let i = 0, j = response.hits.hits.length; i < j; i++ ) {
-					results.push( response.hits.hits[ i ]._source );
-				}
-
-				ReactDOM.render( <RecordsList records={results} />, document.getElementById( "result-details" ) );
 			},
 			error: function( jqXHR, textStatus, errorThrown ) {
 				console.log( jqXHR, textStatus, errorThrown );
